1) Included box funtions in the classes Lattice, Gb, and Bicrystal; 2) Improved the performance of RLLL; 3) Added testGB3das a example of a GB in 3D; 4) updated the examples folder

+++ include/Lattices/RLLL.h
@@ -26,7 +26,7 @@ namespace gbLAB
         
         
         MatrixType B;
-        Eigen::Matrix<long int,Eigen::Dynamic,Eigen::Dynamic> U;
+        Eigen::Matrix<long long int,Eigen::Dynamic,Eigen::Dynamic> U;
         
         /**********************************************************************/
         void update(VectorType& H,
@@ -46,7 +46,7 @@ namespace gbLAB
         /**********************************************************************/
         const MatrixType& reducedBasis() const;
         /**********************************************************************/
-        const Eigen::Matrix<long int,Eigen::Dynamic,Eigen::Dynamic>& unimodularMatrix() const;
+        const Eigen::Matrix<long long int,Eigen::Dynamic,Eigen::Dynamic>& unimodularMatrix() const;
         
     };
